Add files via upload
A simple Blog-site that lets the author add new Blog-entires by accessing the " /compose "-route. Each new entry is added to the "home"-page of the site. Each new post will exists as long as the web page remains active or until the server is shut down, and thus, the site is not supported by a database in its current state. 
The project is largely based on EJS templating and reusing of layouts.   

The project has to run locally on the device or on a server, in order to be operational. 

The following npm-modules are required and used in this project:
- express
- bodyParser
- ejs
- lodash